Function sgx_tservice::rsgx_close_pse_session

pub fn rsgx_close_pse_session() -> SgxError

rsgx_close_pse_session closes a session created by rsgx_create_pse_ session.

Description

An Intel(R) SGX enclave calls rsgx_close_pse_session() when there is no need to request platform service.

Requirements

Header: sgx_tae_service.edl

Library: libsgx_tservice.a

Errors

SGX_ERROR_SERVICE_UNAVAILABLE

The AE service did not respond or the requested service is not supported.

SGX_ERROR_SERVICE_TIMEOUT

A request to the AE service timed out.

SGX_ERROR_UNEXPECTED

Indicates an unexpected error occurs.