Function sgx_tservice::rsgx_get_trusted_time

pub fn rsgx_get_trusted_time(
    current_time: &mut sgx_time_t, 
    time_source_nonce: &mut sgx_time_source_nonce_t
) -> SgxError

rsgx_get_trusted_time gets trusted time from the AE service.

Description

current_time contains time in seconds and time_source_nonce contains nonce associate with the time. The caller should compare time_ source_nonce against the value returned from the previous call of this API if it needs to calculate the time passed between two readings of the Trusted Timer. If the time_source_nonce of the two readings do not match, the difference between the two readings does not necessarily reflect time passed.

The caller should call rsgx_create_pse_session to establish a session with the platform service enclave before calling this API.

Parameters

current_time

Trusted Time Stamp in seconds relative to a reference point. The reference point does not change as long as the time_source_nonce has not changed.

time_source_nonce

A pointer to the buffer that receives the nonce which indicates time source.

Requirements

Header: sgx_tae_service.edl

Library: libsgx_tservice.a

Errors

SGX_ERROR_INVALID_PARAMETER

Any of the pointers is invalid.

SGX_ERROR_AE_SESSION_INVALID

Session is not created or has been closed by architectural enclave service.

SGX_ERROR_SERVICE_UNAVAILABLE

The AE service did not respond or the requested service is not supported.

SGX_ERROR_SERVICE_TIMEOUT

A request to the AE service timed out.

SGX_ERROR_NETWORK_FAILURE

Network connecting or proxy setting issue was encountered.

SGX_ERROR_OUT_OF_MEMORY

Not enough memory is available to complete this operation.

SGX_ERROR_OUT_OF_EPC

There is not enough EPC memory to load one of the Architecture Enclaves needed to complete this operation.

SGX_ERROR_UNEXPECTED

Indicates an unexpected error occurs.