Function sgx_tservice::tae::rsgx_get_ps_sec_prop

pub fn rsgx_get_ps_sec_prop(
    security_property: &mut sgx_ps_sec_prop_desc_t
) -> SgxError

rsgx_get_ps_sec_prop gets a data structure describing the security property of the platform service.

This API is only available in simulation mode.

Description

Gets a data structure that describes the security property of the platform service.

The caller should call rsgx_create_pse_session to establish a session with the platform service enclave before calling this API.

Parameters

security_property

A pointer to the buffer that receives the security property descriptor of the platform service.

Requirements

Header: sgx_tae_service.edl

Library: libsgx_tservice_sim.a (simulation)

Errors

SGX_ERROR_INVALID_PARAMETER

Any of the pointers is invalid.

SGX_ERROR_AE_SESSION_INVALID

Session is not created or has been closed by architectural enclave service.